(refacto): rework backend archi + add insomnia collection
