+++ collection/Insomnia_2021-08-11.json
@@ -1,0 +1,1 @@
+{"_type":"export","__export_format":4,"__export_date":"2021-08-11T08:36:40.641Z","__export_source":"insomnia.desktop.app:v2021.4.1","resources":[{"_id":"req_d178f683a0fb495daa8fdfac46e1aaa2","parentId":"fld_49eecd03c97742c9952b8633a50824d7","modified":1628670182929,"created":1628670140285,"url":"http://localhost:3000/api/purchases/all","name":"get_all_purchases","description":"","method":"GET","body":{},"parameters":[],"headers":[],"authentication":{},"metaSortKey":-1628670140352,"isPrivate":false,"settingStoreCookies":true,"settingSendCookies":true,"settingDisableRenderRequestBody":false,"settingEncodeUrl":true,"settingRebuildPath":true,"settingFollowRedirects":"global","_type":"request"},{"_id":"fld_49eecd03c97742c9952b8633a50824d7","parentId":"wrk_4b5941cbb86c44e1897daa1de25562e7","modified":1628667626814,"created":1628667626814,"name":"purchases","description":"","environment":{},"environmentPropertyOrder":null,"metaSortKey":-1628667626814,"_type":"request_group"},{"_id":"wrk_4b5941cbb86c44e1897daa1de25562e7","parentId":null,"modified":1628667610587,"created":1628667610587,"name":"un_brin_d_aventure","description":"","scope":"collection","_type":"workspace"},{"_id":"req_94cdc8de7243487ba19f5b6193053dc8","parentId":"fld_49eecd03c97742c9952b8633a50824d7","modified":1628670449703,"created":1628670203957,"url":"http://localhost:3000/api/purchases/customer/2","name":"get_customer_purchase","description":"","method":"GET","body":{},"parameters":[],"headers":[],"authentication":{},"metaSortKey":-1628668891952.5,"isPrivate":false,"settingStoreCookies":true,"settingSendCookies":true,"settingDisableRenderRequestBody":false,"settingEncodeUrl":true,"settingRebuildPath":true,"settingFollowRedirects":"global","_type":"request"},{"_id":"req_fe40cf5d53624f7fbe2c6ea7621a656e","parentId":"fld_49eecd03c97742c9952b8633a50824d7","modified":1628670465740,"created":1628670463245,"url":"http://localhost:3000/api/purchases/2","name":"get_purchase_by_id","description":"","method":"GET","body":{},"parameters":[],"headers":[],"authentication":{},"metaSortKey":-1628668267752.75,"isPrivate":false,"settingStoreCookies":true,"settingSendCookies":true,"settingDisableRenderRequestBody":false,"settingEncodeUrl":true,"settingRebuildPath":true,"settingFollowRedirects":"global","_type":"request"},{"_id":"req_c516cc920c9449b189f4a9721092f4f5","parentId":"fld_49eecd03c97742c9952b8633a50824d7","modified":1628670844429,"created":1628670514214,"url":"http://localhost:3000/api/purchases/7","name":"delete_purchase","description":"","method":"DELETE","body":{},"parameters":[],"headers":[],"authentication":{},"metaSortKey":-1628667955652.875,"isPrivate":false,"settingStoreCookies":true,"settingSendCookies":true,"settingDisableRenderRequestBody":false,"settingEncodeUrl":true,"settingRebuildPath":true,"settingFollowRedirects":"global","_type":"request"},{"_id":"req_ab8d44697b4b41d5b4533c7b66c36bbe","parentId":"fld_49eecd03c97742c9952b8633a50824d7","modified":1628670742531,"created":1628670533598,"url":"http://localhost:3000/api/purchases/14","name":"create_purchase","description":"","method":"POST","body":{"mimeType":"application/json","text":"{\n\t\"type\": \"bouquet\",\n\t\"date\": \"11/08/2021\",\n\t\"quantity\": 1,\n\t\"comment\": \"Bonsoir\"\n}"},"parameters":[],"headers":[{"name":"Content-Type","value":"application/json","id":"pair_a0559340c18a433ba18447fecf9e043a"}],"authentication":{},"metaSortKey":-1628667799602.9375,"isPrivate":false,"settingStoreCookies":true,"settingSendCookies":true,"settingDisableRenderRequestBody":false,"settingEncodeUrl":true,"settingRebuildPath":true,"settingFollowRedirects":"global","_type":"request"},{"_id":"req_c7f1c25e5d9f493c8154aee157856167","parentId":"fld_0a2f8a582f734c669c3faa0dba49c8d1","modified":1628670164642,"created":1628667643491,"url":"http://localhost:3000/api/customers/all","name":"get_all_customers","description":"","method":"GET","body":{},"parameters":[],"headers":[],"authentication":{},"metaSortKey":-1628667643553,"isPrivate":false,"settingStoreCookies":true,"settingSendCookies":true,"settingDisableRenderRequestBody":false,"settingEncodeUrl":true,"settingRebuildPath":true,"settingFollowRedirects":"global","_type":"request"},{"_id":"fld_0a2f8a582f734c669c3faa0dba49c8d1","parentId":"wrk_4b5941cbb86c44e1897daa1de25562e7","modified":1628667621546,"created":1628667621546,"name":"customers","description":"","environment":{},"environmentPropertyOrder":null,"metaSortKey":-1628667621546,"_type":"request_group"},{"_id":"req_febfa05386de480e86cc35f43e3b45f9","parentId":"fld_0a2f8a582f734c669c3faa0dba49c8d1","modified":1628668948034,"created":1628668290012,"url":"http://localhost:3000/api/customers/new","name":"create_customer","description":"","method":"POST","body":{"mimeType":"application/json","text":"{\n\t\"name\":\"Lefait\",\n\t\"firstname\": \"Joris\",\n\t\"birthdate\": \"25/10/1992\",\n\t\"address\": \"449 rue clemenceau\",\n\t\"city\": \"Loos\",\n\t\"email\": \"joris.lefait@gmail.com\",\n\t\"phone\": \"0612480174\",\n\t\"comment\": \"\"\n}"},"parameters":[],"headers":[{"name":"Content-Type","value":"application/json","id":"pair_9589b56e845a4eb5833b8add8099f66e"}],"authentication":{},"metaSortKey":-1628667623187,"isPrivate":false,"settingStoreCookies":true,"settingSendCookies":true,"settingDisableRenderRequestBody":false,"settingEncodeUrl":true,"settingRebuildPath":true,"settingFollowRedirects":"global","_type":"request"},{"_id":"req_1fe20697b0664416871fa1aceb2fe5df","parentId":"fld_0a2f8a582f734c669c3faa0dba49c8d1","modified":1628667786892,"created":1628667783150,"url":"http://localhost:3000/api/customers/2","name":"get_customer_by_id","description":"","method":"GET","body":{},"parameters":[],"headers":[],"authentication":{},"metaSortKey":-1628667602821,"isPrivate":false,"settingStoreCookies":true,"settingSendCookies":true,"settingDisableRenderRequestBody":false,"settingEncodeUrl":true,"settingRebuildPath":true,"settingFollowRedirects":"global","_type":"request"},{"_id":"req_e1b86c3290e84f67b7734074c8e7aa79","parentId":"fld_0a2f8a582f734c669c3faa0dba49c8d1","modified":1628668693330,"created":1628667891852,"url":"http://localhost:3000/api/customers/name/Hoa/firstname/city/email","name":"search_customer","description":"","method":"GET","body":{},"parameters":[],"headers":[],"authentication":{},"metaSortKey":-1628667582455,"isPrivate":false,"settingStoreCookies":true,"settingSendCookies":true,"settingDisableRenderRequestBody":false,"settingEncodeUrl":true,"settingRebuildPath":true,"settingFollowRedirects":"global","_type":"request"},{"_id":"env_c98e124e89fc62514dbe89fd560425b7f5b42ad7","parentId":"wrk_4b5941cbb86c44e1897daa1de25562e7","modified":1628667610915,"created":1628667610915,"name":"Base Environment","data":{},"dataPropertyOrder":null,"color":null,"isPrivate":false,"metaSortKey":1628667610915,"_type":"environment"},{"_id":"jar_c98e124e89fc62514dbe89fd560425b7f5b42ad7","parentId":"wrk_4b5941cbb86c44e1897daa1de25562e7","modified":1628667610927,"created":1628667610927,"name":"Default Jar","cookies":[],"_type":"cookie_jar"},{"_id":"spc_a87f4c099d834869a2c784165f24ddc4","parentId":"wrk_4b5941cbb86c44e1897daa1de25562e7","modified":1628667610595,"created":1628667610595,"fileName":"un_brin_d_aventure","contents":"","contentType":"yaml","_type":"api_spec"}]}
